requirements.txt
asyncio==3.4.3
datetime==5.5
base58==2.1.1
aiohttp==3.11.11
requests==2.32.3
solders==0.25.0
solana==0.36.3
logging==0.4.9.6
python-dotenv==1.0.1
httpx==0.28.1
playwright==1.50.0
redis==5.2.1
statistics==1.0.3.5
numpy==2.2.2
